[22/10/2025 01:16h] Refactor App layout and update private routes
